Can now create new detectors

+++ src/app/pages/detectors/detectors.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnInit} from '@angular/core';
 import {Detector} from "../../shared/detector/detector";
 import {DetectorService} from "../../shared/detector/detector.service";
 import {ClientInstance} from "../../shared/client/client-instance";
@@ -12,10 +12,12 @@ export class DetectorsComponent implements OnInit {
 
   detectorList: Detector[] = [];
   currentClientInstance: ClientInstance | null = null;
+  showNewForm: boolean = false;
 
   constructor(
     private detectorService: DetectorService
-  ) { }
+  ) {
+  }
 
   ngOnInit(): void {
     this.loadDetectors();
@@ -34,8 +36,22 @@ export class DetectorsComponent implements OnInit {
     this.loadDetectors();
   }
 
-  removeDetector(detectorId: string) {
-    if (this.detectorService.delete(detectorId).subscribe())
+  removeDetector(detectorId: string | undefined) {
+    if (detectorId === undefined) {
       this.detectorList = this.detectorList.filter(detector => detector.id != detectorId);
+    } else {
+      if (this.detectorService.delete(detectorId).subscribe()) {
+        this.detectorList = this.detectorList.filter(detector => detector.id != detectorId);
+      }
+    }
+  }
+
+  newDetector() {
+    this.showNewForm = true;
+  }
+
+  saveNewDetector(detector: Detector) {
+    this.detectorList.unshift(detector);
+    this.showNewForm = false;
   }
 }
